+++ frontend_tests/node_tests/compose_state.js
@@ -20,3 +20,11 @@ run_test("private_message_recipient", ({override}) => {
     compose_state.private_message_recipient("fred@fred.org");
     assert.equal(compose_state.private_message_recipient(), "fred@fred.org");
 });
+
+run_test("is_topic_field_empty", () => {
+    compose_state.topic("");
+    assert.equal(compose_state.is_topic_field_empty(), true);
+
+    compose_state.topic("foo");
+    assert.equal(compose_state.is_topic_field_empty(), false);
+});