+++ static/js/recent_topics_ui.js
@@ -18,6 +18,7 @@ import * as muted_topics from "./muted_topics";
 import * as narrow from "./narrow";
 import * as narrow_state from "./narrow_state";
 import * as navigate from "./navigate";
+import {page_params} from "./page_params";
 import * as people from "./people";
 import * as recent_senders from "./recent_senders";
 import {get, process_message, topics} from "./recent_topics_data";
@@ -479,6 +480,7 @@ export function update_filters_view() {
         filter_participated: filters.has("participated"),
         filter_unread: filters.has("unread"),
         filter_muted: filters.has("include_muted"),
+        is_spectator: page_params.is_spectator,
     });
     $("#recent_filters_group").html(rendered_filters);
     show_selected_filters();
@@ -590,6 +592,7 @@ export function complete_rerender() {
         filter_unread: filters.has("unread"),
         filter_muted: filters.has("include_muted"),
         search_val: $("#recent_topics_search").val() || "",
+        is_spectator: page_params.is_spectator,
     });
     $("#recent_topics_table").html(rendered_body);
     const $container = $("#recent_topics_table table tbody");