+++ zerver/lib/event_schema.py
@@ -883,7 +883,7 @@ def check_realm_update(
 
     assert "extra_data" not in event.keys()
 
-    if prop in ["notifications_stream_id", "signup_notifications_stream_id"]:
+    if prop in ["notifications_stream_id", "signup_notifications_stream_id", "org_type"]:
         assert isinstance(value, int)
         return
 
@@ -1676,6 +1676,7 @@ group_type = DictType(
         ("id", int),
         ("name", str),
         ("members", ListType(int)),
+        ("subgroups", ListType(int)),
         ("description", str),
         ("is_system_group", bool),
     ]
@@ -1745,6 +1746,28 @@ def check_user_group_update(var_name: str, event: Dict[str, object], field: str)
     assert isinstance(event["data"], dict)
 
     assert set(event["data"].keys()) == {field}
+
+
+user_group_add_subgroups_event = event_dict_type(
+    required_keys=[
+        ("type", Equals("user_group")),
+        ("op", Equals("add_subgroups")),
+        ("group_id", int),
+        ("subgroup_ids", ListType(int)),
+    ]
+)
+check_user_group_add_subgroups = make_checker(user_group_add_subgroups_event)
+
+
+user_group_remove_subgroups_event = event_dict_type(
+    required_keys=[
+        ("type", Equals("user_group")),
+        ("op", Equals("remove_subgroups")),
+        ("group_id", int),
+        ("subgroup_ids", ListType(int)),
+    ]
+)
+check_user_group_remove_subgroups = make_checker(user_group_remove_subgroups_event)
 
 
 user_status_event = event_dict_type(